[Release-1.5.0] 1. reconstruct flow control & failed rpc retry policy; 2. reconstruct fm/lr; 3. add new optimize algorithm; 4. reconstruct default model partition algorithm; 5. add sparse owlqn; 6. optimize servicing; 7. fix a lot of bugs

+++ angel-ps/psf/src/main/java/com/tencent/angel/ml/matrix/psf/update/enhance/map/MapWithIndexFunc.java
@@ -25,5 +25,5 @@ import com.tencent.angel.common.Serialize;
  * as parameter of `MapWithIndex`
  */
 public interface MapWithIndexFunc extends Serialize {
-  double call(int index, double value);
+  double call(long index, double value);
 }
